+++ source/api-client/Energinet.DataHub.MarketParticipant.Client/source/Energinet.DataHub.MarketParticipant.Client/Models/UserIdentityAuditLogEntryDto.cs
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Model
-{
-    public sealed record ErrorResponse
-    {
-        public ErrorResponse(ErrorDescriptor error)
-        {
-            Error = error;
-        }
+using System;
 
-        public ErrorDescriptor Error { get; }
-    }
+namespace Energinet.DataHub.MarketParticipant.Client.Models
+{
+    public sealed record UserIdentityAuditLogEntryDto(
+        Guid UserId,
+        Guid ChangedByUserId,
+        UserIdentityAuditLogField Field,
+        string NewValue,
+        string OldValue,
+        DateTimeOffset Timestamp);
 }