+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Domain/Model/Users/UserIdentityAuditLogField.cs
@@ -10,11 +10,14 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License.using System;
+// limitations under the License.
 
-using System;
-using System.Collections.Generic;
+namespace Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 
-namespace Energinet.DataHub.MarketParticipant.Application.Commands.User;
-
-public sealed record GetAssociatedUserActorsResponse(IEnumerable<Guid> ActorIds);
+public enum UserIdentityAuditLogField
+{
+    FirstName = 1,
+    LastName = 2,
+    PhoneNumber = 3,
+    Status = 4,
+}