+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Domain/Repositories/IUserInviteAuditLogEntryRepository.cs
@@ -18,9 +18,6 @@ using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 
 namespace Energinet.DataHub.MarketParticipant.Domain.Repositories;
 
-/// <summary>
-/// Repository for inserting and querying user roles audit logs
-/// </summary>
 public interface IUserInviteAuditLogEntryRepository
 {
     /// <summary>