+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Tests/Handlers/UpdateUserIdentityHandlerTests.cs
@@ -16,8 +16,10 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
+using Energinet.DataHub.Core.App.Common.Abstractions.Users;
 using Energinet.DataHub.MarketParticipant.Application.Commands.User;
 using Energinet.DataHub.MarketParticipant.Application.Handlers.User;
+using Energinet.DataHub.MarketParticipant.Application.Security;
 using Energinet.DataHub.MarketParticipant.Domain.Exception;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
 using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
@@ -36,8 +38,14 @@ public sealed class UpdateUserIdentityHandlerTests
     public async Task Completes_With_Success()
     {
         // Arrange
+        var userContextFrontendUser = new Mock<IUserContext<FrontendUser>>();
         var userIdentityRepository = new Mock<IUserIdentityRepository>();
         var userRepositoryMock = new Mock<IUserRepository>();
+        var userIdentityAuditLogEntryRepository = new Mock<IUserIdentityAuditLogEntryRepository>();
+
+        userContextFrontendUser
+            .Setup(x => x.CurrentUser)
+            .Returns(new FrontendUser(Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid(), false));
 
         var userIdentityUpdateDto = new UserIdentityUpdateDto("firstName", "lastName", "+45 23232323");
         var validUserId = Guid.NewGuid();
@@ -52,8 +60,10 @@ public sealed class UpdateUserIdentityHandlerTests
             .ReturnsAsync(CreateFakeUserIdentity);
 
         var target = new UpdateUserIdentityHandler(
+            userContextFrontendUser.Object,
             userRepositoryMock.Object,
-            userIdentityRepository.Object);
+            userIdentityRepository.Object,
+            userIdentityAuditLogEntryRepository.Object);
 
         var updateUserIdentityCommand = new UpdateUserIdentityCommand(userIdentityUpdateDto, validUserId);
 
@@ -68,6 +78,11 @@ public sealed class UpdateUserIdentityHandlerTests
     public async Task Invalid_PhoneNumber_ValidationException()
     {
         // Arrange
+        var userContextFrontendUser = new Mock<IUserContext<FrontendUser>>();
+        userContextFrontendUser
+            .Setup(x => x.CurrentUser)
+            .Returns(new FrontendUser(Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid(), false));
+
         var userRepositoryMock = new Mock<IUserRepository>();
         userRepositoryMock
             .Setup(x => x.GetAsync(It.IsAny<UserId>()))
@@ -78,12 +93,16 @@ public sealed class UpdateUserIdentityHandlerTests
             .Setup(x => x.GetAsync(It.IsAny<ExternalUserId>()))
             .ReturnsAsync(CreateFakeUserIdentity());
 
+        var userIdentityAuditLogEntryRepository = new Mock<IUserIdentityAuditLogEntryRepository>();
+
         var userIdentityUpdateDto = new UserIdentityUpdateDto("firstName", "lastName", "+45 invalid");
         var validUserId = Guid.NewGuid();
 
         var target = new UpdateUserIdentityHandler(
+            userContextFrontendUser.Object,
             userRepositoryMock.Object,
-            userIdentityRepository.Object);
+            userIdentityRepository.Object,
+            userIdentityAuditLogEntryRepository.Object);
 
         var updateUserIdentityCommand = new UpdateUserIdentityCommand(userIdentityUpdateDto, validUserId);
 
@@ -95,15 +114,23 @@ public sealed class UpdateUserIdentityHandlerTests
     public async Task UserNotFound_Throws()
     {
         // Arrange
+        var userContextFrontendUser = new Mock<IUserContext<FrontendUser>>();
         var userIdentityRepository = new Mock<IUserIdentityRepository>();
         var userRepositoryMock = new Mock<IUserRepository>();
+        var userIdentityAuditLogEntryRepository = new Mock<IUserIdentityAuditLogEntryRepository>();
+
+        userContextFrontendUser
+            .Setup(x => x.CurrentUser)
+            .Returns(new FrontendUser(Guid.NewGuid(), Guid.NewGuid(), Guid.NewGuid(), false));
 
         var userIdentityUpdateDto = new UserIdentityUpdateDto("firstName", "lastName", "+45 23232324");
         var validUserId = Guid.NewGuid();
 
         var target = new UpdateUserIdentityHandler(
+            userContextFrontendUser.Object,
             userRepositoryMock.Object,
-            userIdentityRepository.Object);
+            userIdentityRepository.Object,
+            userIdentityAuditLogEntryRepository.Object);
 
         var updateUserIdentityCommand = new UpdateUserIdentityCommand(userIdentityUpdateDto, validUserId);